+++ rsconnect/main.py
@@ -169,6 +169,20 @@ def cloud_shinyapps_args(func):
     return wrapper
 
 
+def shinyapps_deploy_args(func):
+    @click.option(
+        "--visibility",
+        "-V",
+        type=click.Choice(["public", "private"]),
+        help="The visibility of the resource being deployed. (shinyapps.io only; must be public (default) or private)",
+    )
+    @functools.wraps(func)
+    def wrapper(*args, **kwargs):
+        return func(*args, **kwargs)
+
+    return wrapper
+
+
 def _passthrough(func):
     @functools.wraps(func)
     def wrapper(*args, **kwargs):
@@ -961,6 +975,7 @@ def deploy_voila(
 @content_args
 @cloud_shinyapps_args
 @click.argument("file", type=click.Path(exists=True, dir_okay=True, file_okay=True))
+@shinyapps_deploy_args
 @cli_exception_handler
 def deploy_manifest(
     name: str,
@@ -977,6 +992,7 @@ def deploy_manifest(
     verbose: bool,
     file: str,
     env_vars: typing.Dict[str, str],
+    visibility: typing.Optional[str],
 ):
     kwargs = locals()
     set_verbosity(verbose)
@@ -1272,6 +1288,7 @@ def generate_deploy_python(app_mode, alias, min_version):
         nargs=-1,
         type=click.Path(exists=True, dir_okay=False, file_okay=True),
     )
+    @shinyapps_deploy_args
     @cli_exception_handler
     def deploy_app(
         name: str,
@@ -1290,6 +1307,7 @@ def generate_deploy_python(app_mode, alias, min_version):
         verbose: bool,
         directory,
         extra_files,
+        visibility: typing.Optional[str],
         env_vars: typing.Dict[str, str],
         image: str,
         account: str = None,